+++ game-core/src/main/java/games/strategy/engine/chat/ChatClient.java
@@ -1,7 +1,7 @@
 package games.strategy.engine.chat;
 
+import org.triplea.domain.data.ChatParticipant;
 import org.triplea.domain.data.UserName;
-import org.triplea.http.client.lobby.chat.ChatParticipant;
 import org.triplea.http.client.lobby.chat.messages.server.ChatMessage;
 import org.triplea.http.client.lobby.chat.messages.server.ChatterList;
 import org.triplea.http.client.lobby.chat.messages.server.StatusUpdate;
