+++ game-core/src/main/java/games/strategy/engine/chat/ChatPlayerListener.java
@@ -1,7 +1,7 @@
 package games.strategy.engine.chat;
 
 import java.util.Collection;
-import org.triplea.http.client.lobby.chat.ChatParticipant;
+import org.triplea.domain.data.ChatParticipant;
 
 /** Callback interface for a components interested in displaying a list of chat participants. */
 public interface ChatPlayerListener {
