+++ game-core/src/main/java/games/strategy/engine/chat/IChatController.java
@@ -3,7 +3,7 @@ package games.strategy.engine.chat;
 import games.strategy.engine.message.IRemote;
 import games.strategy.engine.message.RemoteActionCode;
 import java.util.Collection;
-import org.triplea.http.client.lobby.chat.ChatParticipant;
+import org.triplea.domain.data.ChatParticipant;
 
 /**
  * A central controller of who is in the chat.
