+++ game-core/src/main/java/games/strategy/engine/chat/PlayerChatRenderer.java
@@ -20,8 +20,8 @@ import javax.swing.Icon;
 import javax.swing.ImageIcon;
 import javax.swing.JList;
 import javax.swing.SwingConstants;
+import org.triplea.domain.data.ChatParticipant;
 import org.triplea.domain.data.UserName;
-import org.triplea.http.client.lobby.chat.ChatParticipant;
 
 /**
  * Renders a chat participant in a {@link JList}.
