+++ http-clients/src/main/java/org/triplea/http/client/lobby/chat/messages/server/ChatterList.java
@@ -2,7 +2,7 @@ package org.triplea.http.client.lobby.chat.messages.server;
 
 import java.util.List;
 import lombok.Value;
-import org.triplea.http.client.lobby.chat.ChatParticipant;
+import org.triplea.domain.data.ChatParticipant;
 
 @Value
 public class ChatterList {
