+++ http-server/src/main/java/org/triplea/server/lobby/chat/event/processing/ChatEventProcessor.java
@@ -5,8 +5,8 @@ import java.util.List;
 import java.util.Optional;
 import javax.websocket.Session;
 import lombok.AllArgsConstructor;
+import org.triplea.domain.data.ChatParticipant;
 import org.triplea.domain.data.UserName;
-import org.triplea.http.client.lobby.chat.ChatParticipant;
 import org.triplea.http.client.lobby.chat.messages.client.ChatClientEnvelopeType;
 import org.triplea.http.client.lobby.chat.messages.server.ChatMessage;
 import org.triplea.http.client.lobby.chat.messages.server.ChatServerEnvelopeFactory;
