+++ http-server/src/main/java/org/triplea/server/lobby/chat/event/processing/Chatters.java
@@ -15,8 +15,8 @@ import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
+import org.triplea.domain.data.ChatParticipant;
 import org.triplea.domain.data.UserName;
-import org.triplea.http.client.lobby.chat.ChatParticipant;
 
 @Slf4j
 @AllArgsConstructor
