+++ http-server/src/test/java/org/triplea/server/lobby/chat/event/processing/ChattersTest.java
@@ -25,9 +25,9 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.triplea.domain.data.ChatParticipant;
 import org.triplea.domain.data.PlayerChatId;
 import org.triplea.domain.data.UserName;
-import org.triplea.http.client.lobby.chat.ChatParticipant;
 import org.triplea.server.lobby.chat.event.processing.Chatters.ChatterSession;
 
 @SuppressWarnings("InnerClassMayBeStatic")
